:memo: Add GNU Affero General Public License Version 3
This commit adds the GNU Affero General Public License Version 3 to
the project, a free copyleft license designed to ensure cooperation with
the community in the case of network server software. The license grants
the rights to users to share, distribute, and modify the software, and
requires that modifications made to the code be shared back to the community.

:lipstick: Update footer with Creative Commons license

The footer now includes a Creative Commons Attribution-ShareAlike 4.0 International License. The updated message is more concise and includes the license information at the end of the message, after the software and template credits.